requirements.txt
# We need a "high" cmake version (to be installed in conda env) if we want jaxkd[cuda]
einops<=0.8.1
tqdm<=4.67.1
matplotlib<=3.10.3
numpy<=1.26.4
scikit-learn<=1.7.1
xmipp_metadata
grain<=0.2.11
pandas<=2.1.4
jax[cuda12]<=0.6.0
flax<=0.10.7
tensorflow<=2.19.0
tensorflow-datasets<=4.9.9
# tfp-nightly[jax]<0.27.0
ott-jax<=0.5.1
jaxkd<=0.1.2
orbax<=0.1.9
orbax-checkpoint<=0.11.25
cloudpickle<=3.1.1
dm_pix<=0.4.4
tensorboard<=2.19.0
umap<=0.1.1
pynndescent<=0.5.13
napari[all] >=0.6.0, < 0.7.0
napari-molecule-reader
python_tsp<=0.2.0
torch<=2.0.1
cuml-cu12==25.6.*
umap-learn<0.6.0
mmap_ninja<=0.8.0